# api/swagger/swagger.yaml
basePath: /api/v1
host: localhost:9091
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Velero backup and restore management API with multi-cluster support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Velero API Server
  version: "1.0"
paths:
  /helm/chart_check:
    post:
      consumes:
      - application/json
      description: kubeconfig을 사용하여 특정 헬름 차트 설치 여부 확인
      parameters:
      - description: 헬름 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: 특정 헬름 차트 설치 여부 확인
      tags:
      - helm
  /helm/health:
    get:
      consumes:
      - application/json
      description: kubeconfig을 사용하여 헬름 연결 검증
      parameters:
      - description: 헬름 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: 헬름 연결 확인
      tags:
      - helm
  /helm/install:
    post:
      consumes:
      - application/json
      description: kubeconfig을 사용하여 헬름 차트 설치
      parameters:
      - description: 헬름 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: 헬름 차트 설치
      tags:
      - helm
  /kube/health:
    get:
      consumes:
      - application/json
      description: kubeconfig을 사용하여 쿠버네티스 클러스터 연결 검증
      parameters:
      - description: 쿠버네티스 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: 쿠버네티스 클러스터 연결 확인
      tags:
      - kubernetes
  /kube/pods:
    get:
      consumes:
      - application/json
      description: kubeconfig을 사용하여 쿠버네티스 파드 목록 조회
      parameters:
      - description: 쿠버네티스 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: 쿠버네티스 파드 목록 조회
      tags:
      - kubernetes
  /minio/bucket_check:
    post:
      consumes:
      - application/json
      description: MinioConfig을 사용하여 Minio 버킷 존재 여부 확인 후, 없으면 생성
      parameters:
      - description: Minio 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: Minio 버킷 확인
      tags:
      - minio
  /minio/health:
    get:
      consumes:
      - application/json
      description: MinioConfig을 사용하여 Minio 연결 검증
      parameters:
      - description: Minio 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: Minio 연결 확인
      tags:
      - minio
  /velero/backup-repositories:
    get:
      consumes:
      - application/json
      description: MinioConfig, object을 사용하여 Velero 백업 저장소 목록 조회
      parameters:
      - description: Velero 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: Velero 백업 저장소 목록 확인
      tags:
      - velero
  /velero/backup-storage-locations:
    get:
      consumes:
      - application/json
      description: MinioConfig, object을 사용하여 Velero 백업 스토리지 위치 목록 조회
      parameters:
      - description: Velero 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: Velero 백업 스토리지 위치 목록 확인
      tags:
      - velero
  /velero/backups:
    post:
      consumes:
      - application/json
      description: MinioConfig, object을 사용하여 Velero 백업 목록 조회
      parameters:
      - description: Velero 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: Velero 백업 목록 확인
      tags:
      - velero
  /velero/health:
    get:
      consumes:
      - application/json
      description: object을 사용하여 Velero 연결 검증
      parameters:
      - description: Velero 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: Velero 연결 확인
      tags:
      - velero
  /velero/restores:
    get:
      consumes:
      - application/json
      description: MinioConfig, object을 사용하여 Velero 복구 목록 조회
      parameters:
      - description: Velero 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: Velero 복구 목록 확인
      tags:
      - velero
  /velero/volume-snapshot-locations:
    get:
      consumes:
      - application/json
      description: MinioConfig, object을 사용하여 Velero 볼륨 스냅샷 위치 목록 조회
      parameters:
      - description: Velero 연결에 필요한 값
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 연결 성공
          schema:
            type: object
        "400":
          description: 잘못된 요청
          schema:
            type: object
        "500":
          description: 서버 내부 오류
          schema:
            type: object
        "503":
          description: 서비스 이용 불가
          schema:
            type: object
      summary: Velero 볼륨 스냅샷 위치 목록 확인
      tags:
      - velero
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"